[Close #15] Extend newWindow() to receive optional x y.

+++ menubar-webkit/LDYWebViewWindowController.h
@@ -13,7 +13,7 @@
 @property (weak) IBOutlet WebView *webView;
 
 - (id)initWithURLString:(NSString *)URLString
-                  width:(NSInteger)width
-                 height:(NSInteger)height;
+                  width:(CGFloat)width
+                 height:(CGFloat)height;
 
 @end
